refactor(captp): breakout eager promise resolution forwarding
